+++ src/main/java/domain/card/Card.java
@@ -11,12 +11,12 @@ public class Card {
         this.rank = rank;
     }
 
-    public int getScore() {
-        return rank.getScore();
-    }
-
     public boolean isAce() {
         return rank.isAce();
+    }
+
+    public int getScore() {
+        return rank.getScore();
     }
 
     public String getSymbolValue() {
