+++ src/test/java/domain/ResultTest.java
@@ -1,8 +1,7 @@
-package domain.result;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.Result;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Symbol;
